pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pds-polkastakesphere"
version = "1.0.0"
authors = [{ name = "shamb0" }]
dependencies = [
    "dagster==1.6.6",
    "dagster-dbt==0.22.6",
    "duckdb==0.10.0",
    "dbt-core==1.7.8",
    "dbt-duckdb==1.7.2",
    "dagster-duckdb==0.22.6",
    "ruamel.yaml==0.18.6",
    "pydantic>2",
    "google-cloud-bigquery==3.13.0",
    "google-cloud-bigquery-storage==2.24.0",
    "google-auth==2.24.0",
    "pydantic==2.5.2",
    "db-dtypes==1.1.1",
    "pandas==2.1.3",
    "pyarrow==14.0.2",
    "tqdm==4.65.0",
    "streamlit==1.29.0",
    "matplotlib==3.8.2",
    "pyspark==3.3.2",
    "typing-extensions>=4.6.1"
]

requires-python = ">=3.11, <=3.12"
readme = "README.md"
license = { text = "MIT" }

[project.optional-dependencies]
dev = [
    "click",
    "dagster-webserver",
    "pytest",
    "sqlfluff",
    "sqlfluff-templater-dbt",
    "ruff==0.1.6"
]

tests = [
    "pytest",
    "responses"
]

[tool.setuptools.packages.find]
exclude = ["config","data-vault"]

# [tool.setuptools]
# packages = ["polkastakesphere_dm"]

# [tool.setuptools.package-data]
# "polkastakesphere_dm" = ["../dbt/**"]

# ########################
# ##### DAGSTER
# ########################

[tool.dagster]
module_name = "polkastakesphere_dm.polkastakesphere_dm.definitions"
code_location_name = "polkastakesphere_dm/polkastakesphere_dm"

# ########################
# ##### ruff
# ########################

[tool.ruff]
select = ["E", "F", "W", "I", "C90", "N", "UP", "YTT", "S", "BLE", "FBT", "A", "C4", "T10", "RET", "SIM", "ARG", "PLR", "PIE", "PLC", "PLE", "PLW", "RUF"]
ignore = ["E501", "E203", "E731", "S101", "S106", "A003", "N806", "PLR2004", "COM812", "Q000", "Q003", "E902", "UP007", "S608"]
line-length = 120
fix = true

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.isort]
combine-as-imports = true
force-single-line = true
known-first-party = ["polkastakesphere_dm"]
known-third-party = []
known-local-folder = []

[tool.ruff.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "double"

[tool.ruff.flake8-import-conventions]
aliases = { altair = "alt", bokeh = "bok", matplotlib = "mpl", numpy = "np", pandas = "pd", seaborn = "sns" }

# ########################
# ##### SQLFLUFF
# ########################

[tool.sqlfluff.core]
templater = "dbt"
dialect = "snowflake"
exclude_rules = [
    "aliasing.unique.column",
    "ambiguous.column_count",
    "structure.column_order",
    "structure.using",
    "references.consistent",
    "references.qualification",
]
max_line_length = 100
nocolor = true
# Warn on the use of keywords as identifiers
warnings = "RF04"

[tool.sqlfluff.templater.dbt]
project_dir = "./polkastakesphere_dm/dbt"
profiles_dir = "./polkastakesphere_dm/dbt"

[tool.sqlfluff.templater.jinja]
apply_dbt_builtins = true
load_macros_from_path = "./polkastakesphere_dm/dbt/macros"

[tool.sqlfluff.indentation]
tab_space_size = 4

[tool.sqlfluff.rules.capitalisation.keywords]
capitalisation_policy = "lower"

[tool.sqlfluff.rules.capitalisation.identifiers]
capitalisation_policy = "lower"

[tool.sqlfluff.rules.capitalisation.functions]
extended_capitalisation_policy = "lower"

[tool.sqlfluff.rules.capitalisation.literals]
capitalisation_policy = "lower"

[tool.sqlfluff.rules.capitalisation.types]
extended_capitalisation_policy = "lower"
